agent: extract remoteExecHandler
Set userEventHandler on Agent
Remove shutdownCh, and use a context.

Add an implementation of a net/rpc client

This implementation in a shim that will allow us, in the future, to replace the RPC mechanism with
gRPC. The shim allows not only for changing the underlying implementation, but also for supporting in-place
upgrades.

The context arg is not used yet, but should be once the underlying RPC implementation changes.