proxycfg: pass context around where it is needed
context.Context should never be stored on a struct (as it says in the godoc) because it is easy to
to end up with the wrong context when it is stored.

This change is also in preparation for splitting state into kind-specific handlers so that the
implementation of each kind is grouped together.